add remote user and group options
This role previously used only ansible_ssh_user on the remote
host, which caused quite a bit of pain in controlling which
user should get the rotated key. With this change, the local
(ansible local machine) user is as previously (sshkeys_use)
though the remote user is now specified by sshkeys_remote_user
and sshkeys_remote_group for easier control.

Change-Id: I2ec9b56e58878f63fde625453bfb7ba9b3ca11b3